Fix print, check memory in tests

See merge request eT-program/eT!819

+++ src/memory/memory_manager_class.F90
@@ -226,7 +226,7 @@ contains
 !
       type(memory_manager) :: mem
 !
-      integer, intent(in) :: total 
+      integer(i64), intent(in) :: total 
 !
       character(len=*), intent(in) :: units 
 !
